chore: extract common widgets from wallet link flow
