+++ catalyst_voices/lib/pages/registration/recover/recover_seed_phrase_panel.dart
@@ -1,3 +1,4 @@
+import 'package:catalyst_voices/pages/registration/recover/seed_phrase/account_details_panel.dart';
 import 'package:catalyst_voices/pages/registration/recover/seed_phrase/seed_phrase_input_panel.dart';
 import 'package:catalyst_voices/pages/registration/recover/seed_phrase/seed_phrase_instructions_panel.dart';
 import 'package:catalyst_voices/pages/registration/widgets/placeholder_panel.dart';
@@ -18,7 +19,7 @@ class RecoverSeedPhrasePanel extends StatelessWidget {
       RecoverSeedPhraseStage.seedPhraseInstructions =>
         const SeedPhraseInstructionsPanel(),
       RecoverSeedPhraseStage.seedPhrase => const SeedPhraseInputPanel(),
-      RecoverSeedPhraseStage.linkedWallet => const PlaceholderPanel(),
+      RecoverSeedPhraseStage.accountDetails => const AccountDetailsPanel(),
       RecoverSeedPhraseStage.unlockPasswordInstructions =>
         const PlaceholderPanel(),
       RecoverSeedPhraseStage.unlockPassword => const PlaceholderPanel(),
